SYNC-1554 Group batch events by site so that they can be canceled if it is deactivated

+++ modules/util/sync-engine/src/main/java/com/liferay/sync/engine/documentlibrary/handler/MoveToTrashHandler.java
@@ -31,6 +31,10 @@ public class MoveToTrashHandler extends BaseJSONHandler {
 	public void processResponse(String response) throws Exception {
 		SyncFile syncFile = getLocalSyncFile();
 
+		if (syncFile == null) {
+			return;
+		}
+
 		SyncFileService.deleteSyncFile(syncFile);
 	}
